Fix NullPointerException when calling _cleanup when listFiles returns null.
Since this is usually caused by a permissions issue, a WARN level log event is issued to indicate the problem, and the _cleanup call returns a 500 to indicate a server-side error.

Closes #83.

+++ src/main/java/com/github/rnewson/couchdb/lucene/LuceneServlet.java
@@ -91,14 +91,21 @@ public final class LuceneServlet extends HttpServlet {
 				}
 			}
 
-			// Delete all indexes except the keepers.
-			for (final File dir : DatabaseIndexer.uuidDir(root, db.getUuid())
-					.listFiles()) {
-				if (!viewKeep.contains(dir.getName())) {
-					LOG.info("Cleaning old index at " + dir);
-					FileUtils.deleteDirectory(dir);
-				}
-			}
+            // Delete all indexes except the keepers.
+            final File[] dirs = DatabaseIndexer.uuidDir(root, db.getUuid()).listFiles();
+            if (dirs == null) {
+                LOG.warn(DatabaseIndexer.uuidDir(root, db.getUuid())
+                         + " is not a directory or could not be read.");
+                ServletUtils.sendJSONError(req, resp, 500, "index_dir_perms");
+                return;
+            } else {
+                for (final File dir : dirs) {
+                    if (!viewKeep.contains(dir.getName())) {
+                        LOG.info("Cleaning old index at " + dir);
+                        FileUtils.deleteDirectory(dir);
+                    }
+                }
+            }
 		}
 
 		// Delete all directories except the keepers.
